Patched a segmentation fault
When adding a filter to an existing MP, the solution segmentation faulted after row 474, unless a Log::verbose() was issued. I suspect, that this is due to that the result is not filled by the mysql_store_result (despite the doc. say it would). Hence, this is a nasty 'fix'. Pending proper solution.  I compared to the 'boot' approach, and they seem similar in execution, that worked this failed...